# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SakanMatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /paypal/create-subscription:
    post:
      operationId: createSubscription
      tags: [paypal]
      summary: Create a PayPal subscription for premium
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSubscriptionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Owners only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /paypal/activate-subscription:
    post:
      operationId: activateSubscription
      tags: [paypal]
      summary: Activate a PayPal subscription after approval
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateSubscriptionRequest"
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivateSubscriptionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /paypal/cancel-subscription:
    post:
      operationId: cancelSubscription
      tags: [paypal]
      summary: Cancel the current user's active PayPal subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: No active subscription to cancel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    CreateSubscriptionResponse:
      type: object
      properties:
        subscriptionID:
          type: string
        approvalUrl:
          type: string
          nullable: true
      required:
        - subscriptionID
    ActivateSubscriptionRequest:
      type: object
      properties:
        subscriptionID:
          type: string
      required:
        - subscriptionID
    ActivateSubscriptionResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
          nullable: true
      required:
        - status
...