# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SakanMatch API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const signupBodyPasswordMin = 6;

export const SignupBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
  name: zod.string().optional(),
  role: zod.enum(["owner", "seeker"]),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string().nullish(),
    role: zod.enum(["owner", "seeker"]),
    isPremium: zod.boolean(),
    premiumActivatedAt: zod.string().nullish(),
    premiumSource: zod.string().nullish(),
    subscriptionStatus: zod.string().nullish(),
    createdAt: zod.string().optional(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string().nullish(),
  role: zod.enum(["owner", "seeker"]),
  isPremium: zod.boolean(),
  premiumActivatedAt: zod.string().nullish(),
  premiumSource: zod.string().nullish(),
  subscriptionStatus: zod.string().nullish(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Get all listings
 */
export const GetListingsQueryParams = zod.object({
  city: zod.coerce.string().optional(),
  minPrice: zod.coerce.number().optional(),
  maxPrice: zod.coerce.number().optional(),
});

export const GetListingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  price: zod.number(),
  city: zod.string(),
  images: zod.array(zod.string()).optional(),
  ownerId: zod.number(),
  ownerEmail: zod.string().optional(),
  ownerName: zod.string().nullish(),
  isFeatured: zod.boolean().optional(),
  viewCount: zod.number().nullish(),
  contactClickCount: zod.number().nullish(),
  requestCount: zod.number().nullish(),
  createdAt: zod.string().optional(),
});
export const GetListingsResponse = zod.array(GetListingsResponseItem);

/**
 * @summary Create a listing
 */
export const CreateListingBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  city: zod.string(),
  images: zod.array(zod.string()).optional(),
});

/**
 * @summary Get current user listings
 */
export const GetMyListingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  price: zod.number(),
  city: zod.string(),
  images: zod.array(zod.string()).optional(),
  ownerId: zod.number(),
  ownerEmail: zod.string().optional(),
  ownerName: zod.string().nullish(),
  isFeatured: zod.boolean().optional(),
  viewCount: zod.number().nullish(),
  contactClickCount: zod.number().nullish(),
  requestCount: zod.number().nullish(),
  createdAt: zod.string().optional(),
});
export const GetMyListingsResponse = zod.array(GetMyListingsResponseItem);

/**
 * @summary Record a contact click for a listing
 */
export const RecordContactClickParams = zod.object({
  id: zod.coerce.number(),
});

export const RecordContactClickResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get a listing by ID
 */
export const GetListingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetListingResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  price: zod.number(),
  city: zod.string(),
  images: zod.array(zod.string()).optional(),
  ownerId: zod.number(),
  ownerEmail: zod.string().optional(),
  ownerName: zod.string().nullish(),
  isFeatured: zod.boolean().optional(),
  viewCount: zod.number().nullish(),
  contactClickCount: zod.number().nullish(),
  requestCount: zod.number().nullish(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Delete a listing
 */
export const DeleteListingParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteListingResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Create a PayPal subscription for premium
 */
export const CreateSubscriptionResponse = zod.object({
  subscriptionID: zod.string(),
  approvalUrl: zod.string().nullish(),
});

/**
 * @summary Activate a PayPal subscription after approval
 */
export const ActivateSubscriptionBody = zod.object({
  subscriptionID: zod.string(),
});

export const ActivateSubscriptionResponse = zod.object({
  status: zod.string(),
  message: zod.string().nullish(),
});

/**
 * @summary Cancel the current user's active PayPal subscription
 */
export const CancelSubscriptionResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary PayPal webhook handler
 */
export const PaypalWebhookResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get premium status of current user
 */
export const GetPremiumStatusResponse = zod.object({
  isPremium: zod.boolean(),
  premiumActivatedAt: zod.string().nullish(),
  premiumSource: zod.string().nullish(),
  subscriptionStatus: zod.string().nullish(),
  paypalSubscriptionId: zod.string().nullish(),
  lastPaymentAt: zod.string().nullish(),
});

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});